chore: 🤖 Switching from old queue to the new
